ACompactPull (Pull request from ErikEJ to fix SQL Compact perf issue: CodePlex 857)
The actual fix submitted is a one line change to the SQL Compact provider to add a cast as described in the bug. I verified that we have functional coverage for this code path and then also did a little bit of refactoring so that I could add unit tests.