Fix for #902: General performance regression introduced by changeset f15fb3b6614e
Problem: In fixing #849, perf regressed because certain metadata collections never go read-only and so the fast-lookup-by-identity dictionary was no longer being built.

Solution: Revert the fix for #849 and change MetadataCollection so that it invalidates the internal cache when, either:
1) An item is removed.
2) An item identity is changed (currently we only need to worry about EntitySet and EdmMember).

The identity cache is now built lazily so that it is only built/rebuilt the first time it is needed.

Verification: David has tested this change with various regressed perf tests and determined that we are now either as fast or faster than before the regression.