WhoMovedMyWensleydale (Add back EdmFunctionAttribute and EntityFunctions)
For EF6 these classes were moved to the top-level namespace and renamed to DbFunctions and DbFunctionAttribute. However, this has caused more confusion than expected with people not being able to find the moved classes. This change puts back versions of these classes with their original names so that people can easily find them. The classes with the old names basically just delegate to the new classes and are also marked as obsolete to help move people to the new classes.