FlyingTheNest (Enable Code First to use nested types)
This change allows nested entity, complex, and enum types to be mapped when using Code First. This change builds on the o/c mapping changes for Code First that have already be made. It adds support for nested types by taking the CLR nested type name and transforming it into an o-space model namespace and name such that the o-space namespace consists of the CLR namespace combined with the outer type name(s) and the o-space type name is the simple CLR type name. For example, "My.Namespace.OuterType+NestedType1+NestedType2" will become "My.Namespace.OuterType.NestedType1" "NestedType2" in o-space.

Note that the c-space namespace name is still the same for all types in the model and the c-space type name is still taken from the simple type of the CLR type. This means that only a single type with a given simple name can be used in a single model. It is however fine to have types in the same assembly that have the same simple name so long as they are used in different models.

This change also introduces a new model builder version for EF6. Nested types will be discovered with this model builder version or with the 'Latest' version, but will not be discovered with earlier model builder versions. Explicitly mapping nested types is allowed when using older versions since there is nothing in the schema that limits this.

EdmModel has been changed to use model builder version instead of schema version because it must now handle models that build differently even though the generated schema version will be the same.