JustBecauseYouAreAHighCharacter... (Regression test for CodePlex 190)
This was a bug about using a TVF in ESQL with an SSDL namespace that has high byte characters. This appears to work in the EF6 code base, either because it was fixed since the bug was filed (probably before the move to Git) or because it was actually a test issue with Taupo rather than a real bug.