MunroPineappleSociety (Fix for CodePlex 478: model differ crash with spatial property)
This was a bug with type conversion checking in ColumnModel where the list of primitive types had not been updated to include spatial types. Fixing this allows the model differ to now succeed, although such a conversion will still fail at the SQL level because it needs a rebuild.

I added a functional test for the scenario in the bug and also for changing spatial column names. There doesn't appear to be many more alterations that can be done to spatial columns. I also added a unit test that will fail if new primitive types are added without updating the ColumnModel class.