GoneWithTheWinmd - Fix for EF fails when a winmd assembly is referenced
EF sometimes tries loading all referenced assemblies (e.g. when looking for
pre-compiled views). However with Win8 the main assembly may have a
reference to a winmd assembly. Trying to load winmd assemblies with
Assembly.Load causes a FileLoadException which currently is unhandled. The
fix is just to catch the FileLoadException and swallow it as we already do
for FileNotFoundException.

Work item: 609