TheUpsideDownShow_Season2 (Prevent database initializers from creating a database if a Migrations configuration is discovered)
Fix for CodePlex 529: Upgrade to EF6 breaks migrations prior to upgrade.

The repro of this bug involves this:
1. Enabling Migrations and running some migrations
2. Deleting or clearing the database entirely
3. Creating the database again using the default (non-Migrations) database initializer
4. Attempting to update the database to its previous state using Migrations

Step 3 is the critical step since it creates a database using an automatic migration rather than using the migrations that were setup in step 1.

One solution is to to discover the DbMigrationsConfiguration if possible and use it to update the database if found. This means that the explicit Migrations set up by the user in step 1 are run again and everything ends up in a consistent state.

This solution was rejected by the team room as being too unexpected. Instead, the initializer will do nothing in this case leaving the database uncreated and still in a state that Update-Database can be run.

Note that if the MigrationsConiguration cannot be found (e.g. because it is in a different assembly) then it is still possible to get the exception reported in the bug, but I don't know that there is much that can be done about that, or if it is worth doing anything. I suspect that the scenario in this bug is most likely to happen by mistake when people are playing rather than in an actual working Migrations development environment.