TurnItUpToEleven (Allow model differ to handle MaxLength="MAX" in SSDL)
Fix for CodePlex 726: Differ: Handle MaxLength="MAX" in SSDL

SSDL allows a property to be set as MaxLength="Max" which is equivalent to leaving out the MaxLength attribute. The SQL Server provider never does this (it leaves out MaxLength) but other providers can behave differently. This change fixes the parsing of MaxLength in SSDL so that it treats Max the same as not set.