ShroedingersFunctionalTests - fixes the intermittent test failures in the C.I. environment
The C.I. machine is showing intermittent test failures after changeset e95ca12e5ae6 was committed. It seems to be related to the timing at which certain tests are executed. Some test make use of the DynamicTypeDescriptionConfiguration<T> class which helps to dynamically inject new attributes to the types used by the test models. Since we are now caching the attributes we discover for any given property (see class AttributeProvider) the Dispose() method of class DynamicTypeDescriptionConfiguration<T> is explicitly clearing the attribute cache.

In very rare cases (coincidentally never in our development machines, but sometimes in the C.I. machine) the cache is not cleared in time before a new test is executed, one that can particularly be affected by the attribute injection of a previous test. In order to alleviate this situation, we need to explicitly clear the cache before any of these potentially-affected tests run. Given xUnit's design, the best place to do this is in the Dispose() method of the class that hosts the [Fact] tests that can affect other tests. All of these tests reside in class DataAnnotationScenarioTests.