"TotoIveGotAFeelingWereNotInTheFrameworkAnymore (CodePlex 859: Removing System.Data.Entity reference on EF6 package install)
Adds to the existing .NET code that is called when the package is installed to also find and remove a reference to System.Data.Entity.dll if present. This could be done entirely in PowerShell, but using the .NET code allows the code to be more easily tested and maintained. In addition to the test I also did manual package installs for both projects with an SDE reference and projects without an SDE reference.