SlashDotDashDotSlashDotComDotComDotComDotDotDotDotCom (Add validation for Migrations directory to avoid downstream problems)
This came out of CodePlex 948 where a leading slash in MigrationsDirectory causes entire project to be copied into a sub-folder. This was because MigrationsDirectory must be a relative path to a directory under the VS project, but we weren't ever validating that this was the case.

After discussion in the team room we decided to throw a "known exception" for this so that the Package Manager Console will display a nice message without a stack trace. However, it may be useful to have a stack trace for this or other known exceptions, so now if you turn on -Verbose we will spit out a stack trace even for known exceptions.