ThatsAWrap (Adding tests that use wrapping providers.)
In an attempt to understand the root cause of http://entityframework.codeplex.com/workitem/959 I created some tests that use wrapping ADO.NET and EF providers. I was unable to repro the issue, but this is useful coverage to have anyway.

These tests revealed one minor issue whereby we would use the query cache that is contained on the MetadataWorkspace if the same model (e.g. for HistoryContext) is used once with a non-wrapped provider and then again in the same app domain with a wrapped provider that uses the same provider invariant name. It would be very hard to get into this sitation in the wild since it requires dynamic mutation of the configuration while EF is running, but I fixed the bug anyway in order to make the tests pass.