HaveABreakHaveAKitKat (Fix for net40 build break)
The problem was that the provider rows were being cached by the provider row finder, which meant that the change to the ADO.NET provider table was not picked up by the test. This only showed on .NET 4 because the row finder is only used there. On .NET 4.5 the newly exposed property on DbConnection is used instead.

Since the row finder has traditionally been a source of errors I did some refactoring to allow it to be tested and debugged on .NET 4.5 even though it is not typically used on .NET 4.5.