ChiliCarneCon (Updating FunctionalTestsConfiguration to avoid test threading issue)
We have a test (Configurations_can_be_pushed_and_popped_from_multiple_threads_concurrently) that causes instance FunctionalTestsConfiguration to be locked from multiple threads at the same time. This could fail when accessing the static OriginalConnectionFactories list which was a non-thread safe collection. The fix is to manipulate the collection in a thread-safe way. Because of the way the tests use this collection I don't believe that there needs to any protection around which value wins in the race condition where two threads attempt to add at the same time. The only test that will do this is the above mentioned test, and all threads will be adding the same value anyway.