DbqWrapWithCurlyFries (Allow LINQ to Entities to unwrap DbQueries into ObjectQueries)
Fix for CodePlex 5: DbQuery wrapping LINQ provider fails to recognize DbQuery that needs to be rewritten as ObjectQuery

Prior to EF6 there was no way for LINQ to Entities to recognize the DbQuery, and there was no way for the DbQuery provider to incept and generate an ObjectQuery when the IQueryable was passed diectly to LINQ to Entities.

I looked at fixing this by removing the wrapping provider, but this is a significant task with some risk due to:
- The different ways that generic and non-generic ObjectQueries are handled compared to the way that generic and non-generic DbQueries are handled.
- The use of the DbQuery/InternalQuery structuring for testing.

Since the fix for the bug is simple and seems low risk I chose to go this way instead. We should sill consider removing the wrapping provider, but probably not for EF6

I also moved DbFunctionScenarios out of the transitional project. The only internal method it was using was .Each.