fixing Issue 771 - DbContext with open connection does not work with TransactionScope.
Problem is that we are creating multiple connections to the same database during context initialization. If user starts local transaction - using (new TransactionScope()) then all these connections are part of the transaction. If on top of that user provides us with a opened connection that is part of this transaction then this causes promotion to distributed transaction. The fix is to suppress transactions for all querying operations we do during context initialization. If context has to perform some CUD operations, it may still fail, but if its up to date and just being re-initialized it will now work, which was not the case before.

CR: AVickers, LaJones