TimeGotTheTimeTickTickTickinInMyHead (Add first-class support for CommandTimeout in DbContext)
It was previously possible to set the CommandTimeout for most operations by dropping down to ObjectContext. This change introduces a property of context.Database.CommandTimeout which will be used for all operations of the context. This includes DDL and implicit Migrations operations. Static (i.e. non-context) DDL methods on database obviously don't use this timeout.

DbMigrationsConfiguration already contained a CommandTimeout property used when running Migrations. This is now also used for history context operations.