PrivacyPolicyUpdated (Enable use of unmapped properties with private setters on unmapped base classes)
The scenario here is an entity like so:

public class MyBase
{
    public int Id { get; private set; }
}

public class MyEntity : MyBase
{
}

Code First would not allow this to be mapped and furthermore if it could be mapped then core EF would not be able to write to the property. The reason for this is that PropertyInfo.CanWrite returns false if the PropertyInfo is obtained from the derived class. In addition, methods that use the setter also don't work--for example SetValue, and GetSetMethod.

The fix is to look for the PropertyInfo for the same property on the property's DeclaringType amd see whether it is writable. If it is, then this is used throughout the stack for writing.

I attempted to functional test all the areas that I could think of where EF writes to properties and I think I have caught them all.