ShareAndEnjoy (Code First: Allow subclasses to map property to same database column with TPH)
EDM allows properties in different subclasses of a TPH hierarchy to map to the same database column and thereby "share" that column. Code First was building this mapping mostly correctly except that it was not fixing up the mappings to all point to the same column. This change fixes that.

Note that since the column can now be configured from multiple properties all configurations of the column must match or an exception is thrown. Configuration can be applied from any one property or from multiple properties so long as the configuration does not conflict.