MigrationsAreEmigrating (Allow migrations and context to live in different projects)
This change adds a new option of ContextProjectName for Enable-Migrations that, when specified, makes Migrations look in the given project for the DbContext for which Migrations should be enabled. Note that both the migrations project and the context project must have the Entity-Framework package installed, and ultimately the migrations project must have a reference to the context project, although this is not needed to actually run the Enable-Migrations command.