IsTrivialRenameReallyTrivialAfterAll - fix for workitem 563 - Linq to EF Join throws "Index was out of range" after upgrade from VS2010 to VS2012 http://entityframework.codeplex.com/workitem/563
Changes to the product code had already been code reviewed, but I'm sending another code review for EF6 since I'm now adding a few functional tests to cover the scenario that exposed this issue.

The code included in this changeset makes use of a new method called IsTrivialRename() which in turn tries to access newExpresion.Arguments[0] and newExpression.Arguments[1] without checking if these arguments exist in the collection. This is what's causing this regression.

The issue is easily fixed by checking whether newExpression.Arguments[...] and newExpression.Parameters[...] contain enough values (2) to perform the checks that the method intends to do.