SheSellsPowerShellsByTheSeaShore (Add PowerShell commands to make provider configuration easier)
This change is a prerequisite to some up-coming changes that will allow a DbProviderServices implementation to provide multiple provider services without needing to register each one separately. These commands make it easy for provider-writers to register their DbProviderServices in the config file on NuGet package install. A future change will also allow the newly registered provider to be made the default.

The code makes use of the Migrations PowerShell infrastructure. I also changed the existing non-Migrations PowerShell interaction to use the same infrastructure and be called in the same way using PowerShell commands. These things are:
- Initializing the entityFramework section
- Doing SQL Express/LocalDb detection and setting the connection factory default
- Removing System.Data.Entity references
- Setting the default connection factory in the SQL Compact provider package.

We can consider trying to remove setting the default connection factory and just using the provider default, but this is complicated due to the LocalDb/Express detection and due to
the chance of breaking existing apps. Therefore this is not changed at this time. It is however now easy for any other provider to set a default connection factory on package
install.