GoingOldSkool (Replace IHistoryContextFactory interface with HistoryContextFactory delegate)
IHistoryContextFactory existed as a way to register a service that could create customized HistoryContext instances. However, it was a pain to have to create a class for this when just a delegate would be fine. The problems with using Func<DbConnection, bool, string> is that it is not very helpful for Intellisense or very specific/useful for defining a service. So instead this change introduces a delegate type that provides help and is a reasonable type to define a service for.