CSDLMeansCSpace - fixing a model validation rule
ValidateReferentialConstraint rule is checking whether the properties on the dependent side
are key properties for IAs. According to the comment this rule should be enforced on CSpace
models. The check was missing and after we enabled validation for SSpace models the rule was
enforced also on SSpace models. This breaks the designer since the schema views contain models
that trigger this condition. Also note that this condition could only be triggered for v1
models but all schema views (even for reverse engineering v3 concepts) are v1.