WonderboltsAcademy (Fix for build/test perf regression)
OneProviderToRuleThemAll caused the amount of time to run build.cmd to increase substantially. The problem was (as suspected) the provider loading code. The fix is two-fold. First, the expensive Reflection code to create the provider instances has been moved into the AppConfig class and cached there so that for a given config it only happens once. Second, when Migrations is running on behalf of a DbContext it needs to make sure that the correct configuration is used, which makes use of the push/pop configuration functionality. But if Migrations is being used with the default config for the app domain and for an assembly that is already known about, then there is no need to push a new configuration and so this is now avoided.