ShareTheLove (Ensure inherited shared columns are handled correctly)
This is an additional fix for the recent change to allow properties to share columns in a TPH table. There are two parts to the fix. First, for shared properties that are inherited by a further subclass the code was attempting to remove a column that had already been removed. This worked in release builds but asserted that the item being removed was not present when using a debug build. This part of the fix is to check before removing.

Second, removing columns means that column names that were temporarily uniquified no longer need to be uniquified. This is handled by SetReadOnly. However, SetReadOnly was not notifying the underlying collection that identities had changed. This was fine when the identity dictionary had not yet been built, but in some cases the dictionary has already been built, which then caused the exception reported in the bug. The fix is to notify the underlying collection when an identity changes as part of SetReadOnly.