Pull request for 965: Consider adding DbSet.AddRange API
This change was made by UnaiZorrilla with a small amount of additional cleanup by ajcvickers.

From bug: Invoking DbSet.Add multiple times when numerous separate entities need to be added can be very slow as each call to Add will trigger DetectChanges. We provide guidance on how to turn off automatic change detection to improve performance of this scenario but that is not very discoverable. AddRange should be very easy to implement and test and would be a much more convenient and discoverable way to add many entities.