CollectionRecollection - fixing writing FunctionImport return parameters and relaxing validation for collections
There is a bug in the designer where return parameters for functions may not be created correctly
(i.e. the return type is a RowType instead a collection of RowType). This was inadvertently compensated
during serialization by always writing the CollectionType header for function RowType return types.
This change fixes this - now we are writing the actual type.
Since we are now handling Collections we need to relax some validation rules. Collection types - similarly to
row types - are created by EF and the user cannot change names or namespaces so the corresponding
rules should not report errors.