GACAttack (Make by-convention loading of SQL Server provider use type name with version and public key token)
This fixes the issue where the SQL Server provider would fail to load when the EF assemblies are GACed and no provider is registered.

Since EF is never GACed automatically and bin-deployment is strongly preferred over GACing there is no change made to the entries added in the config file by default. This means that if EF is GACed and the config files are being used, then they may need to be manually updated to contain extra type information. By not doing this for everyone we improve the experience for the vast majority who are not GACing at the expense of an extra step for those who are.