MockYourDoppelgänger (Make DbSet suitable for creating test doubles)
As per design meeting discussion (http://entityframework.codeplex.com/wikipage?title=Design%20Meeting%20Notes%20-%20May%2016%2c%202013) this change makes the DbSet classes and their related classes suitable for creating test doubles by adding protected constructors and making methods virtual as appropriate.

This is somewhat fiddly because these classes are IQueryables, involve generics and non-generics, and already have a class hierarchy. I played with various implementation options including creating a null object for InternalSet/InternalQuery or adding additional types into the hierarchy but ultimately these became more messy than just testing for null InternalSet/InternalQuery, so that's where it ended up.

Methods and properties on the doubles must be implemented since there is no context or other backing classes to provide functionality. To make this clear and avoid hard-to-find issues most members throw if they are used but not implemented in the test double. The exceptions to this are Include, AsNoTracking, and AsStreaming where default no-op implementations are reasonable and useful. Also, ToString never throws, but reverts to the default .NET behavior.

Testing was done as functional tests to replicate the environment that an application developer would be using when creating test doubles--that is, without any access to internals. The test matrix includes tests for generic and non-generic DbSets crossed with completely unimplemented test doubles (mostly to test exceptions), test doubles with in-memory implementations, and test doubles created dynamically with Moq.