InceptionInsideInterception  (Ensure DbContext is in interception context for s-space command trees)
The problem here was the way EntityProviderServices is used to create the c-space command tree and then goes through the public ADO.NET API to call back into itself to then use the store provider services to create the o-space tree. In going through the public surface (CreateCommandDefinition) the interception context was lost. The fix is to add a internal virtual to EntityProviderServices that allows the interception context to be preserved when going through the "normal" path inside EF while still preserving the public surface, EntityClient behavior, and regular provider services behavior.