ZoningIn (Add timezone info to DbCommandLogger output per design meeting decision)
The code uses DateTimeOffset.Now instead of DateTime.Now. The tests didn't need to change because they are written in such a way as to not depend on the format of the DateTime output, but I did manual verification that the output now looks like:

SELECT TOP (2)
  [c].[Id] AS [Id],
  [c].[Title] AS [Title]
  FROM [dbo].[Blogs] AS [c]
  -- Executing at 5/28/2013 4:10:48 PM -07:00
  -- Completed in 1 ms with result: SqlDataReader

I also fixed CodePlex 1160 "LogCommand_logs_command_text_and_parameters test failed with Hungarian culture" while I was in the code. The fix is to use current culture to format the decimal value in the expected string for the Assert.