TheHumberBridge (Bridging the gap between DbConfiguration and Interception)
This change creates a bridge between interception and the code-based configuration. In general, interceptors can be added and removed dynamically at any time, whereas code-based configuration is intended to be unchanging for the application. However, if you want to setup an interceptor that is added at application start and never removed then it makes sense to think of this as code-based configuration. There is now a DbConfiguration.AddInterceptor method to allow this.

The implementation required that a GetServices method be added to IDbDependencyResolver. This method returns all services that satisfy a given type and key combination. This means that many IDbInterceptors can be registered and returned by either adding them in DbConfiguration or by any other IDbDependencyResolver mechanism, including an IoC container. These interceptors are read once as EF is initializing and are registered just as if they had been registered directly on Interception.