LookDontTouch (Ensure DbContextInfo does not cause database initializer to run)
DbContextInfo is used when attempting to get information about a context at design time. The context created by CreateInstance of DbContextInfo should not allow database initialization to happen.

Since DbContextInfo now ensures that the initializer will not run in means that DbMigrator does not need special code for this, so that code is removed.