WhateverPropertyISayIAmThatsWhatImNot (CodePlex 1215: Fix private setter code to handle indexer properties correctly)
CanWriteExtended is used to check whether or not a base type contains a private setter for the property being considered. Indexers are not supported and would be filtered out anyway, but sometimes this code is called before such filtering happens. Therefor the code needs to handle indexer properties and return false.