SerialSerialization (CodePlex 1107: Allow ISafeSerialaztionData exceptions to be serialized over and over)
The issue is that due to the extralinguistic nature of deserialization none of the constructors for the exception are being called and hence SerializeObjectState event is not being hooked up on the deserialized exception. This has been fixed by making sure that the event is hooked up on the deserialized object inside the CompleteDeserialization method.