Runtime: Two fixes:
1) #1136 - Make update command trees publicly constructible

Solution: Made the ctors public and also added ToString() to DbCommandTree. ToString calls Print() which renders a textual version of the tree - highly useful when debugging/testing trees.

2) #1190 - Code First Stored Procs :: exception thrown when using stored procs to map entity with self reference and 1-1 reference to another entity with non-int key

Problem: When generating update trees for entities involved in one-to-one associations using IAs, we first need to change the relationship state to deleted before setting it to added.
Solution: Make the additional call to ChangeRelationshipState when this case is detected. In order to make the call, we need to fabricate a second stub principal entity and that entity needs a key different from the stub entity we are using to set the relationship state to added (otherwise state manager validation will fail), so I had to add the ability to generate fake key values for the stub principal. I also noticed that we were not handling spatial keys correctly and so I fixed that too.

Work Item: 1136, 1190