WhoGotTheFunc (CodePlex 396: Handle closures containing DbSet/DbQuery members)
The code we previously had for this was broken in that it only handled DbQuery/DbSet members that were typed as IQueryable. However, this was already almost fixed in EF6 with the change to handle funcletization of DbQueries directly in LINQ to Entities. The one tweak needed to make this case work and remove the need for the translation in the wrapping provider was to make sure that the expression type of the ObjectQuery expression created by the funcletizer is ObjectQuery or a derivation. The code was previously just using the type of the expression passed in, which was okay when the expression coming in was an ObjectQuery, but not okay if it was a DbQuery that had been converted to an ObjectQuery.